require('pull-ws/client') without including `http`
export duplex in separate file
